Add net.version rpc endpoint
