+++ trinity/rpc/modules/__init__.py
@@ -2,5 +2,5 @@ from .main import RPCModule  # noqa: F401
 
 from .eth import Eth  # noqa: F401
 from .evm import EVM  # noqa: F401
-
+from .net import Net  # noqa: F401
 from .web3 import Web3  # noqa: F401
